added date filter as temporary workaround for New_GLand_2011 header parsing into bogus dates.
Not sure why this is only happening with NewGLand_2011 -- other
fluxall_YYYY.xls files have the same header format.  The better
way to deal with this is to properly parse the fluxall files in
the first place.  This will cease to be a problem when I change
the earlier parts of the pipeline away from Excel, so this
bandaid is temporary.