implemented output as dataset in UNM_flux_processor
trying to track down problem with negative wind directions for
grassland 2009.  The output is getting garbled -- GLand_FluxAll_2009
is ending up with some values in its wind direction column that are
not the wind directions that matlab calculated.  Rather than track
down what is going on in the existing code I am going to begin
implementing an easier-to-follow dataset-based data structure chain
for shepherding data through processing.