implemented "blocking files":
"Blocking files" are a temporary file with a line or two of text.
Matlab can be made to not continue execution while the blocking file
exists.  Useful for running system calls; Matlab on our windows 7 will
not wait for the system call to finish.  This apparently varies across
windows versions.